ml-server/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
Pillow
librosa
tensorflow
torch
torchvision
scikit-learn
